renamed language.ts to LanguageStore
