+++ src/store/index.ts
@@ -1,14 +1,13 @@
 import Vue from 'vue'
 import Vuex from 'vuex'
 
-import languages from './languages'
+import languageStore from './languageStore'
 //import nativelanguages from './nativelanguages';
 
 Vue.use(Vuex)
 
 export default new Vuex.Store({
   modules: {
-    languages,
-   // nativelanguages
+    languageStore
   }
 })
